update jaxtype hints

+++ src/jimgw/prior.py
@@ -94,8 +94,8 @@ class Prior(Distribution):
 
 @jaxtyped(typechecker=typechecker)
 class Uniform(Prior):
-    xmin: Float = 0.0
-    xmax: Float = 1.0
+    xmin: float = 0.0
+    xmax: float = 1.0
 
     def __repr__(self):
         return f"Uniform(xmin={self.xmin}, xmax={self.xmax})"
@@ -149,8 +149,8 @@ class Uniform(Prior):
 
 @jaxtyped(typechecker=typechecker)
 class Unconstrained_Uniform(Prior):
-    xmin: Float = 0.0
-    xmax: Float = 1.0
+    xmin: float = 0.0
+    xmax: float = 1.0
 
     def __repr__(self):
         return f"Unconstrained_Uniform(xmin={self.xmin}, xmax={self.xmax})"
@@ -400,19 +400,19 @@ class PowerLaw(Prior):
     p(x) ~ x^{\alpha}
     """
 
-    xmin: Float = 0.0
-    xmax: Float = 1.0
-    alpha: Float = 0.0
-    normalization: Float = 1.0
+    xmin: float = 0.0
+    xmax: float = 1.0
+    alpha: float = 0.0
+    normalization: float = 1.0
 
     def __repr__(self):
         return f"Powerlaw(xmin={self.xmin}, xmax={self.xmax}, alpha={self.alpha}, naming={self.naming})"
 
     def __init__(
         self,
-        xmin: Float,
-        xmax: Float,
-        alpha: Union[Int, Float],
+        xmin: float,
+        xmax: float,
+        alpha: Union[Int, float],
         naming: list[str],
         transforms: dict[str, tuple[str, Callable]] = {},
         **kwargs,
@@ -425,7 +425,7 @@ class PowerLaw(Prior):
         self.xmin = xmin
         self.alpha = alpha
         if alpha == -1:
-            self.normalization = 1.0 / jnp.log(self.xmax / self.xmin)
+            self.normalization = float(1.0 / jnp.log(self.xmax / self.xmin))
         else:
             self.normalization = (1 + self.alpha) / (
                 self.xmax ** (1 + self.alpha) - self.xmin ** (1 + self.alpha)
@@ -479,10 +479,10 @@ class Exponential(Prior):
     p(x) ~ exp(\alpha x)
     """
 
-    xmin: Float = 0.0
-    xmax: Float = jnp.inf
-    alpha: Float = -1.0
-    normalization: Float = 1.0
+    xmin: float = 0.0
+    xmax: float = jnp.inf
+    alpha: float = -1.0
+    normalization: float = 1.0
 
     def __repr__(self):
         return f"Exponential(xmin={self.xmin}, xmax={self.xmax}, alpha={self.alpha}, naming={self.naming})"
